Moves Hit to components
